// packages/sdk/src/database.types.d.ts
export type Json = string | number | boolean | null | {
    [key: string]: Json | undefined;
} | Json[];
export type Database = {
    public: {
        Tables: {
            [_ in never]: never;
        };
        Views: {
// ...
            v_due_pms: {
                Row: {
                    asset_id: string | null;
                    asset_name: string | null;
                    completion_count: number | null;
                    created_at: string | null;
                    description: string | null;
                    id: string | null;
                    is_active: boolean | null;
                    last_completed_at: string | null;
                    next_due_date: string | null;
                    template_id: string | null;
                    template_name: string | null;
                    tenant_id: string | null;
                    title: string | null;
                    trigger_config: Json | null;
                    trigger_type: string | null;
                    updated_at: string | null;
                };
                Relationships: [
                    {
                        foreignKeyName: "pm_schedules_asset_id_fkey";
                        columns: ["asset_id"];
                        isOneToOne: false;
                        referencedRelation: "v_assets";
                        referencedColumns: ["id"];
                    },
                    {
                        foreignKeyName: "pm_schedules_asset_id_fkey";
                        columns: ["asset_id"];
                        isOneToOne: false;
                        referencedRelation: "v_mobile_assets";
                        referencedColumns: ["id"];
                    },
                    {
                        foreignKeyName: "pm_schedules_template_id_fkey";
                        columns: ["template_id"];
                        isOneToOne: false;
                        referencedRelation: "v_pm_templates";
                        referencedColumns: ["id"];
                    },
                    {
                        foreignKeyName: "pm_schedules_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "mv_tenant_overview";
                        referencedColumns: ["tenant_id"];
                    },
                    {
                        foreignKeyName: "pm_schedules_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_dashboard_metrics";
                        referencedColumns: ["tenant_id"];
                    },
                    {
                        foreignKeyName: "pm_schedules_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_portfolio_overview";
                        referencedColumns: ["tenant_id"];
                    },
                    {
                        foreignKeyName: "pm_schedules_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_tenants";
                        referencedColumns: ["id"];
                    },
                    {
                        foreignKeyName: "pm_schedules_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_tenants_overview";
                        referencedColumns: ["tenant_id"];
                    }
                ];
            };
// ...
            v_pm_template_checklist_items: {
                Row: {
                    created_at: string | null;
                    description: string | null;
                    display_order: number | null;
                    id: string | null;
                    required: boolean | null;
                    template_id: string | null;
                    tenant_id: string | null;
                };
                Relationships: [
                    {
                        foreignKeyName: "pm_template_checklist_items_template_id_fkey";
                        columns: ["template_id"];
                        isOneToOne: false;
                        referencedRelation: "v_pm_templates";
                        referencedColumns: ["id"];
                    },
                    {
                        foreignKeyName: "pm_templates_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "mv_tenant_overview";
                        referencedColumns: ["tenant_id"];
                    },
                    {
                        foreignKeyName: "pm_templates_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_dashboard_metrics";
                        referencedColumns: ["tenant_id"];
                    },
                    {
                        foreignKeyName: "pm_templates_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_portfolio_overview";
                        referencedColumns: ["tenant_id"];
                    },
                    {
                        foreignKeyName: "pm_templates_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_tenants";
                        referencedColumns: ["id"];
                    },
                    {
                        foreignKeyName: "pm_templates_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_tenants_overview";
                        referencedColumns: ["tenant_id"];
                    }
                ];
            };
            v_pm_templates: {
                Row: {
                    created_at: string | null;
                    description: string | null;
                    id: string | null;
                    is_system: boolean | null;
                    name: string | null;
                    tenant_id: string | null;
                    trigger_config: Json | null;
                    trigger_type: string | null;
                    updated_at: string | null;
                    wo_description: string | null;
                    wo_estimated_hours: number | null;
                    wo_priority: string | null;
                    wo_priority_entity_type: string | null;
                    wo_title: string | null;
                };
                Insert: {
                    created_at?: string | null;
                    description?: string | null;
                    id?: string | null;
                    is_system?: boolean | null;
                    name?: string | null;
                    tenant_id?: string | null;
                    trigger_config?: Json | null;
                    trigger_type?: string | null;
                    updated_at?: string | null;
                    wo_description?: string | null;
                    wo_estimated_hours?: number | null;
                    wo_priority?: string | null;
                    wo_priority_entity_type?: string | null;
                    wo_title?: string | null;
                };
                Update: {
                    created_at?: string | null;
                    description?: string | null;
                    id?: string | null;
                    is_system?: boolean | null;
                    name?: string | null;
                    tenant_id?: string | null;
                    trigger_config?: Json | null;
                    trigger_type?: string | null;
                    updated_at?: string | null;
                    wo_description?: string | null;
                    wo_estimated_hours?: number | null;
                    wo_priority?: string | null;
                    wo_priority_entity_type?: string | null;
                    wo_title?: string | null;
                };
                Relationships: [
                    {
                        foreignKeyName: "pm_templates_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "mv_tenant_overview";
                        referencedColumns: ["tenant_id"];
                    },
                    {
                        foreignKeyName: "pm_templates_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_dashboard_metrics";
                        referencedColumns: ["tenant_id"];
                    },
                    {
                        foreignKeyName: "pm_templates_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_portfolio_overview";
                        referencedColumns: ["tenant_id"];
                    },
                    {
                        foreignKeyName: "pm_templates_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_tenants";
                        referencedColumns: ["id"];
                    },
                    {
                        foreignKeyName: "pm_templates_tenant_id_fkey";
                        columns: ["tenant_id"];
                        isOneToOne: false;
                        referencedRelation: "v_tenants_overview";
                        referencedColumns: ["tenant_id"];
                    },
                    {
                        foreignKeyName: "pm_templates_wo_priority_fk";
                        columns: ["tenant_id", "wo_priority_entity_type", "wo_priority"];
                        isOneToOne: false;
                        referencedRelation: "v_priority_catalogs";
                        referencedColumns: ["tenant_id", "entity_type", "key"];
                    }
                ];
            };
// ...
        };
// ...
    };
};